+++ cypress/locators/pod/index.js
@@ -10,5 +10,6 @@ export const podFooter = () => podComponent().find(POD_FOOTER);
 export const podContent = () => podComponent().find(POD_CONTENT);
 export const podSubTitle = () => cy.get(POD_SUBTITLE);
 export const podDescription = () => cy.get(POD_DESCRIPTION);
+export const podEdit = () => podComponent().find(POD_EDIT_ICON);
 
 export const podEditIframe = () => cy.iFrame(POD_DATA_COMPONENT).find(POD_EDIT_ICON);