+++ src/components/menu-list/index.js
@@ -1,2 +1,0 @@
-export { default as MenuList } from "./menu-list.component";
-export { default as MenuListItem } from "./menu-list-item/menu-list-item.component";