+++ src/components/multi-step-wizard/multi-step-wizard.style.js
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+const StyledMultiStepWizard = styled.div`
+  width: 100%;
+`;
+
+const StyledMultiStepWizardContent = styled.div`
+  width: 90%;
+  padding-left: 5%;
+  padding-top: 10px;
+`;
+
+export { StyledMultiStepWizard, StyledMultiStepWizardContent };